Added inheritance
